move PageRank stuff into its own package
