+++ algo/src/main/java/org/neo4j/graphalgo/impl/pagerank/PageRankAlgorithm.java
@@ -16,17 +16,15 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.graphalgo.impl;
+package org.neo4j.graphalgo.impl.pagerank;
 
 import org.neo4j.graphalgo.api.Graph;
 import org.neo4j.graphalgo.api.HugeGraph;
-import org.neo4j.graphalgo.api.RelationshipWeights;
-import org.neo4j.graphalgo.core.heavyweight.HeavyGraph;
 import org.neo4j.graphalgo.core.utils.paged.AllocationTracker;
+import org.neo4j.graphalgo.impl.Algorithm;
 
 import java.util.concurrent.ExecutorService;
 import java.util.stream.LongStream;
-import java.util.stream.Stream;
 
 public interface PageRankAlgorithm {
