+++ algo/src/main/java/org/neo4j/graphalgo/impl/pagerank/PageRankResult.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.graphalgo.impl;
+package org.neo4j.graphalgo.impl.pagerank;
 
 import org.neo4j.graphalgo.core.write.Exporter;
